refactor: the isolate board is a whiteboard
The name had already split, and the wrong half was mine. The saved object
has been called a whiteboard since it reached the schema, the bindings,
the project menu and the board's own header; the mode stayed `isolate` in
sixty places across seventeen files. Every commit from here would have
deepened that.

`isolate` described the original gesture -- lift the AOI off the map --
and that is now a small part of it. Areas are arranged, named, ordered and
compared on this surface, and work is STARTED on it. Whiteboard describes
what the thing is rather than how it began, and it unifies the mode with
the document, which is right: the mode IS the document being edited, and
"what is the difference between isolate mode and a saved whiteboard" had
no answer.

Not `workspace`, which is already the docked layout. Two things under one
name is the failure this session spent four fixes on.

The document keeps the name and the interior keeps its prefix: Whiteboard
is the saved thing and what a user reads; `board*` is the surface, the
scene, the memory, the column, the layout. Naming the component
`Whiteboard` too would have created exactly the ambiguity being removed --
`Whiteboard` is already the store's type.

The button's own words went with it. It said "Isolate the analysis from
the map" and offered "Nothing to isolate: no raster on the map", both of
which describe a mode that no longer exists: it opens on an area with no
raster at all now, which is the whole point of the last three commits.

Mechanical and checked as such: tsc clean, fourteen headless checks
re-pointed and passing, check:contrast clean, every module served, and the
`three` chunk still reached only from inside the renamed directory. The
one remaining "isolated" is a chart comment about dots.
